requirements.txt
alembic==1.16.1
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
flasgger==0.9.5
Flask==2.3.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
mistune==3.1.3
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.7.0
pytest==7.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
repoze.lru==0.7
requests==2.32.4
Routes==2.5.1
rpds-py==0.25.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.46.2
swagger_ui_bundle==1.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==2.3.7
uvicorn
python-jose[cryptography]